refactor: merge unit tests of sglang and vllm engines  (#514)
* merge tests

* fix docstring

* add test

* fix format

* shorter ctx len for test

* add adv norm in grpo test

* update test to use local path

* resource cleanup in tests

* fix vllm pp

* fix

* .

* .

* .

* add assertion

+++ areal/core/remote_inf_engine.py
@@ -606,7 +606,8 @@ class RemoteInfEngine:
             # Update LoRA state if this was a LoRA update
             if meta.use_lora:
                 self.lora_initialized = True
-            shutil.rmtree(meta.path, ignore_errors=True)
+            if meta.clear_checkpoint_after_load:
+                shutil.rmtree(meta.path, ignore_errors=True)
 
         fut.add_done_callback(callback)
